# perf(nexmark): optimize nexmark generation