--- a/src/utils/nexmark/Cargo.toml
+++ b/src/utils/nexmark/Cargo.toml
@@ -6,6 +6,5 @@ description = "Nexmark data stream generator."
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
-chrono = { version = "0.4", default-features = false, features = ["clock", "std"] }
 rand = { version = "0.8", features = ["small_rng"] }
 serde = { version = "1", features = ["derive"], optional = true }