--- a/src/utils/nexmark/src/event.rs
+++ b/src/utils/nexmark/src/event.rs
@@ -39,7 +39,7 @@ use rand::{Rng, SeedableRng};
 use serde::{Deserialize, Serialize};
 
 use crate::config::{GeneratorConfig, CHANNEL_NUMBER};
-use crate::utils::{milli_ts_to_timestamp_string, NexmarkRng};
+use crate::utils::NexmarkRng;
 
 type Id = usize;
 
@@ -113,7 +113,7 @@ pub struct Person {
     /// One of several US states as a two-letter string.
     pub state: String,
     /// A millisecond timestamp for the event origin.
-    pub date_time: String,
+    pub date_time: u64,
     /// Extra information
     pub extra: String,
 }
@@ -146,7 +146,7 @@ impl Person {
             credit_card,
             city,
             state,
-            date_time: milli_ts_to_timestamp_string(time),
+            date_time: time,
             extra,
         }
     }
@@ -182,9 +182,9 @@ pub struct Auction {
     /// The minimum price for the auction to succeed.
     pub reserve: usize,
     /// A millisecond timestamp for the event origin.
-    pub date_time: String,
+    pub date_time: u64,
     /// A UNIX epoch timestamp for the expiration date of the auction.
-    pub expires: String,
+    pub expires: u64,
     /// The ID of the person that created this auction.
     pub seller: Id,
     /// The ID of the category this auction belongs to.
@@ -207,9 +207,7 @@ impl Auction {
         let initial_bid = rng.gen_price();
 
         let reserve = initial_bid + rng.gen_price();
-        let date_time = milli_ts_to_timestamp_string(time);
-        let expires =
-            milli_ts_to_timestamp_string(time + Self::next_length(events_so_far, rng, time, cfg));
+        let expires = time + Self::next_length(events_so_far, rng, time, cfg);
         let mut seller = if rng.gen_range(0..cfg.hot_seller_ratio) > 0 {
             (Person::last_id(id, cfg) / cfg.hot_seller_ratio_2) * cfg.hot_seller_ratio_2
         } else {
@@ -227,7 +225,7 @@ impl Auction {
             description,
             initial_bid,
             reserve,
-            date_time,
+            date_time: time,
             expires,
             seller,
             category,
@@ -290,7 +288,7 @@ pub struct Bid {
     /// The url of this bid
     pub url: String,
     /// A millisecond timestamp for the event origin.
-    pub date_time: String,
+    pub date_time: u64,
     /// Extra information
     pub extra: String,
 }
@@ -327,7 +325,7 @@ impl Bid {
             auction: auction + nex.first_auction_id,
             bidder: bidder + nex.first_person_id,
             price,
-            date_time: milli_ts_to_timestamp_string(time),
+            date_time: time,
             channel,
             url,
             extra,
@@ -364,7 +362,7 @@ mod tests {
                 credit_card: "4355 0142 3460 9324".into(),
                 city: "boise".into(),
                 state: "ca".into(),
-                date_time: "2015-07-15 00:00:00".into(),
+                date_time: 1436918400000,
                 extra: "cllnesmssnthtljklifqbqcyhcjwiuoaudxxwcnnwgmsmwgqelplzyckqzuoaitfpxubgpkjtqjhktelmbskvjkxrhziyowxibbgnqneuaiazqduhkynvgeisbxtknbxmqmzbgnptlrcyigjginataks".into(),
             })
         );