--- a/src/utils/nexmark/src/utils.rs
+++ b/src/utils/nexmark/src/utils.rs
@@ -21,7 +21,7 @@ const MIN_STRING_LENGTH: usize = 3;
 
 pub trait NexmarkRng {
     fn gen_string(&mut self, max: usize) -> String;
-    fn gen_string_with_delimiter(&mut self, max: usize, delimiter: &str) -> String;
+    fn gen_string_with_delimiter(&mut self, max: usize, delimiter: char) -> String;
     fn gen_exact_string(&mut self, length: usize) -> String;
     fn gen_next_extra(&mut self, current_size: usize, desired_average_size: usize) -> String;
     fn gen_price(&mut self) -> usize;
@@ -29,43 +29,26 @@ pub trait NexmarkRng {
 
 impl NexmarkRng for SmallRng {
     fn gen_string(&mut self, max: usize) -> String {
-        self.gen_string_with_delimiter(max, " ")
+        self.gen_string_with_delimiter(max, ' ')
     }
 
-    fn gen_string_with_delimiter(&mut self, max: usize, delimiter: &str) -> String {
+    fn gen_string_with_delimiter(&mut self, max: usize, delimiter: char) -> String {
         let len = self.gen_range(MIN_STRING_LENGTH..max);
-        String::from(
-            (0..len)
-                .map(|_| {
-                    if self.gen_range(0..13) == 0 {
-                        delimiter.to_string()
-                    } else {
-                        ::std::char::from_u32('a' as u32 + self.gen_range(0..26))
-                            .unwrap()
-                            .to_string()
-                    }
-                })
-                .collect::<Vec<String>>()
-                .join("")
-                .trim(),
-        )
+        (0..len)
+            .map(|_| {
+                if self.gen_range(0..13) == 0 {
+                    delimiter
+                } else {
+                    self.gen_range(b'a'..=b'z') as char
+                }
+            })
+            .collect()
     }
 
     fn gen_exact_string(&mut self, length: usize) -> String {
-        let mut s = String::new();
-        let mut rnd = 0;
-        let mut n = 0;
-        for _ in 0..length {
-            if n == 0 {
-                rnd = self.gen();
-                n = 6; // log_26(2^31)
-            }
-            let c = ::std::char::from_u32('a' as u32 + (rnd % 26)).unwrap();
-            s.push(c);
-            rnd /= 26;
-            n -= 1;
-        }
-        s
+        (0..length)
+            .map(|_| self.gen_range(b'a'..=b'z') as char)
+            .collect()
     }
 
     fn gen_next_extra(&mut self, current_size: usize, desired_average_size: usize) -> String {
@@ -73,7 +56,7 @@ impl NexmarkRng for SmallRng {
             return String::new();
         }
         let desired_average_size = desired_average_size - current_size;
-        let delta = (desired_average_size as f64 * 0.2).round() as usize;
+        let delta = desired_average_size / 5;
         let min_size = desired_average_size - delta;
         let desired_size = min_size
             + if delta == 0 {
@@ -91,9 +74,9 @@ impl NexmarkRng for SmallRng {
 
 pub fn get_base_url(seed: u64) -> String {
     let mut rng = SmallRng::seed_from_u64(seed);
-    let id0 = rng.gen_string_with_delimiter(5, "_");
-    let id1 = rng.gen_string_with_delimiter(5, "_");
-    let id2 = rng.gen_string_with_delimiter(5, "_");
+    let id0 = rng.gen_string_with_delimiter(5, '_');
+    let id1 = rng.gen_string_with_delimiter(5, '_');
+    let id2 = rng.gen_string_with_delimiter(5, '_');
     format!(
         "https://www.nexmark.com/{}/{}/{}/item.htm?query=1",
         id0, id1, id2