I hereby agree to the terms of the [Singularity Data, Inc. Contributor License Agreement](https://gist.github.com/skyzh/0663682a70b0edde7ae991492f2314cb#file-s9y_cla).

## What's changed and what's your intention?

This PR optimizes nexmark event generation by:
1. store the timestamp in u64 (ms since UNIX epoch). avoid formatting to string and parsing back.
2. reduce string allocation in the generator.

And adds a micro-benchmark for it.

This improves the performance 2x in micro-benchmark `cd utils/nexmark && cargo bench`:

```
nexmark_generate_event  time:   [442.05 ns 445.75 ns 449.98 ns]
                        change: [-67.326% -67.075% -66.837%] (p = 0.00 < 0.05)
                        Performance has improved.
```

In the simulation test `cargo sslt --release -- ./e2e_test/source/nexmark_endless.slt`,
this reduces the execution time from 2.0s to 1.3s, and the percentage of overhead from 32% to 23%.

## Checklist

- [x] I have written necessary rustdoc comments
- [x] I have added necessary unit tests and integration tests
- [x] All checks passed in `./risedev check` (or alias, `./risedev c`)

## Refer to a related PR or issue link (optional)
#4555