@wangrunji0408 2022-11-15
@BugenZhao mentioned another point to optimize: the nexmark generator generates all 3 kinds of events, but only one kind will be kept at a time. So there will be 3x generation, of which 2x is wasted. I'll fix this issue in a future PR.